requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cffi==1.17.1
pycparser==2.22
